Email: - Implemented the sending of mails with and without attachments 	> Usage w/o attachments: MailManager.sendMail(to, subject, body, mime-type) 	> Usage with attachments: MailManager.sendAttachmentMail(to, subject, body, mime-type, attachment, attachment-mime-type, attachment-filename)
The mail from which these emails are send is "bitsei.wa@gmail.com", created on purpose for this scope by myself.
All the settings are in the MailManager.properties file, DO NOT CHANGE THEM!
I was forced by gmail settings to enable two-step authentication, If someone wants or needs the access to the email write me a pm.